% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_dose_time.R
\name{adjust_dose_time}
\alias{adjust_dose_time}
\title{adjust the dose time}
\usage{
adjust_dose_time(.x, .adjusted, .next_dose_time)
}
\arguments{
\item{.x}{time}

\item{.adjusted}{whether adjustment had already occured}

\item{.next_dose_time}{vector of value(s) to adjust to}
}
\description{
adjust the dose time
}
\details{
given a row derived from a missing dose, the time would have already been
properly adjusted, hence should not tweak, however, to maintain consistency
new ADDL rows time-of-day should match the next dose time-of-day, so 
must adjust to reflect
eg given a next dose time of 7:33
eg record came from missing would already be set to 7:33 - don't adjust
dose record from prior valid dose set to 10:12 --> adjust to actually dose at 7:33
}
