% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_functions.R
\name{append.CovT}
\alias{append.CovT}
\title{Append time-varying covariates to an event data frame}
\usage{
append.CovT(event.df, ...)
}
\arguments{
\item{event.df}{The event dataframe to append to}

\item{...}{Time-varying covariate dataframes to append}
}
\value{
A tbl_df object with the appended data
}
\description{
Append time-varying covariates to an event data frame
}
\details{
If there are mismatched column names the resulting column will have
NA entries for the records sourced from the input data frame that lacks those columns.
For instance, if A.df and B.df input: \code{append.events(A.df, B.df)}, with A.df having
columns {ID, TIME, x3} and B.df having columns {ID, TIME, x4}, the resulting dataframe
will have columsn {ID, TIME, x3, x4}.  Some of the entries for x3 and x4 will be NA.

This is expected behavior for merging time-varying covariates, because the covariates
are not likely to have been measured at the same time as dosing and obersations.
Specifically, dose and observation records will likely have NA values for the covariate
columns.

NA values in columns should be addressed after the merge process is completed.

All input data frames must have ID and TIME columns.
}
