% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_functions.R
\name{append.events}
\alias{append.events}
\title{Append event data frames}
\usage{
append.events(...)
}
\arguments{
\item{...}{Data frames to append}
}
\value{
A tbl_df object with the appended data
}
\description{
Append event data frames
}
\details{
If there are mismatched column names the resulting column will have
NA entries for the records sourced from the input data frame that lacks those columns.
For instance, if A.df and B.df input: \code{append.events(A.df, B.df)}, with A.df having
columns {ID, TIME, x3} and B.df having columns {ID, TIME, x4}, the resulting dataframe
will have columsn {ID, TIME, x3, x4}.  Some of the entries for x3 and x4 will be NA.

NA values in columns should be addressed after the merge process is completed.

All input data frames should have ID and TIME columns.
}
