% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.r
\name{as.data.frame.domain.metadata}
\alias{as.data.frame.domain.metadata}
\title{Format a metadata object as a dataframe}
\usage{
\method{as.data.frame}{domain.metadata}(met, type = "md")
}
\arguments{
\item{met}{A metadata object}

\item{type}{md for markddown formatting of multiline cells}
}
\value{
a data_frame (tbl_df) object
}
\description{
Format a metadata object as a dataframe
}
