% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{bind_rows}
\alias{bind_rows}
\title{bind rows for dataframes with units}
\usage{
bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{dataframes to bind together}
}
\description{
bind rows for dataframes with units
}
\examples{
library(dplyr)
library(PMDatR)
th1 <- set_units_from_list(Theoph, list(Wt = "kg")) \%>\%
    mutate(WTunit="kg")
th2 <- set_units_from_list(Theoph, list(Wt = "lb")) \%>\%
    mutate(WTunit = "lb")

#the WTFLG groups are both scaled to kg
bind_rows(th1, th2) \%>\%
    distinct(Subject, WTunit, .keep_all = TRUE)
}
