% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_text.R
\name{blank_to_missing}
\alias{blank_to_missing}
\title{Convert whitespace to a missing value}
\usage{
blank_to_missing(x, missing_value = NA)
}
\arguments{
\item{x}{A character vector}

\item{missing_value}{Defaults to NA, but can be used to convert all whitespace values to a single character value}
}
\value{
the modified character vector
}
\description{
Convert whitespace to a missing value
}
\details{
Given a character vector, this function replaces entries that contain *only* whitespace with the missing_value
provided.  Entries such as empty strings "", any number of spaces "  ", "     ", tab ("\\t"), newline ("\\n"), and
carriage return ("\\r") are converted to the missing_value.
}
\examples{
blank_to_missing(c("A","B","C"," ", "   ", "\\t", " he", "hi\\t",123,"\\n", "two\\nlines", ""))
}
