% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_numeric.R
\name{boxcox}
\alias{boxcox}
\title{Applies Box-Cox transformation with known lambda}
\usage{
boxcox(x, lambda = 0)
}
\arguments{
\item{x}{Values to transform}

\item{lambda}{Power parameter}
}
\value{
the transformed values.
}
\description{
Applies Box-Cox transformation with known lambda
}
\examples{
boxcox(c(1,2,3))
boxcox(1:3, lambda=1)
}
