% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\name{cfb_abs}
\alias{cfb_abs}
\title{absolute change from baseline}
\usage{
cfb_abs(.df, .col, ..., groups = "ID", .name = NULL,
  .name_fn = function(x) {     paste0("CFB_", x) }, .agg_fn = NULL)
}
\arguments{
\item{.df}{data frame}

\item{.col}{column name (unquoted)}

\item{...}{filter functions (unquoted) to establiash baseline row, defaults to first row per group}

\item{groups}{vector of groups, defaults to "ID"}

\item{.name}{name of created column (quoted).  Defaults to CFB_<.col>.}

\item{.name_fn}{function to apply formatting to the given colun name}

\item{.agg_fn}{function to calculate an aggregate statistic for the baseline value}
}
\description{
absolute change from baseline
}
