% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\name{cfb_ratio}
\alias{cfb_ratio}
\title{ratio change from baseline}
\usage{
cfb_ratio(.df, .col, ..., groups = "ID", .name = NULL,
  .name_fn = function(x) {     paste0("CFBR_", x) }, digits = 3,
  .agg_fn = NULL)
}
\arguments{
\item{.df}{data frame}

\item{.col}{column name (unquoted)}

\item{...}{filter functions (unquoted) to establiash baseline row, defaults to first row per group}

\item{groups}{vector of groups, defaults to "ID"}

\item{.name}{name of created column (quoted).  Defaults to CFBR_<.col>.}

\item{.name_fn}{function to apply formatting to the given colun name}

\item{digits}{number of digits to round final value to}

\item{.agg_fn}{function to calculate an aggregate statistic for the baseline value}
}
\description{
ratio change from baseline
}
