% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_utilities.R
\name{combine_rows_by_tol}
\alias{combine_rows_by_tol}
\title{Combine data frame rows by tolerance on a value}
\usage{
combine_rows_by_tol(df, values = "TIME", tol = 0, ...)
}
\arguments{
\item{df}{A dataframe object}

\item{values}{the quoted name of the column on which to combine rows}

\item{tol}{The tolerance within value in which to combine rows}

\item{...}{Summary function to adjust or create columns within the combined row}
}
\value{
the modified dataframe
}
\description{
Combine data frame rows by tolerance on a value
}
\examples{
set.seed(12345)
vals = runif(100,0,100)
vals = vals[order(vals)]
data1 = data_frame(TIME=vals, SEX=sample(x=0:1,size = 100,replace=TRUE), AMT=100)
# make groups where start is within tol=0.5
combine_rows_by_tol(data1, values="TIME", tol=0.5, AMT=sum(AMT))
}
