% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mappings.R
\name{compare_mappings}
\alias{compare_mappings}
\title{compare_mappings}
\usage{
compare_mappings(map1, map2)
}
\arguments{
\item{map1}{First mappings list}

\item{map2}{Second mappings list}
}
\value{
a dataframe listing the differences
}
\description{
compare_mappings
}
\details{
Compares two domain mappings lists and gives a data frame
with the differences.
}
