% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{contains_one_of}
\alias{contains_one_of}
\title{Check if a value matches multiple regular expressions}
\usage{
contains_one_of(x, ..., ignore.case = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{...}{unquoted or quoted regular expressions}

\item{ignore.case}{A logical.  If FALSE (default) case is ignored in the comparison}
}
\value{
The vector x is checked against each regular expression in the array provided.  Any match
results in TRUE for the element
}
\description{
Check if a value matches multiple regular expressions
}
\examples{
nams=c("George Washington", "Washington DC", "King George")
contains_one_of(nams, Washington, DC)
contains_one_of(nams, ing)
contains_one_of(nams, "^Wash.*$")
}
