% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{contains_one_of_}
\alias{contains_one_of_}
\title{Check if a value matches multiple regular expressions}
\usage{
contains_one_of_(x, regexps, ignore.case = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{regexps}{A vector of regular expressions}

\item{ignore.case}{A logical.  If FALSE (default) case is ignored in the comparison}
}
\value{
A logical vector of length x
}
\description{
Check if a value matches multiple regular expressions
}
\details{
The vector x is checked against each regular expression in the array provided.  Any match
results in TRUE for the element
}
\examples{
nams=c("George Washington", "Washington DC", "King George")
contains_one_of_(nams, c("Washington","DC"))
}
