% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.pmunits}
\alias{convert.pmunits}
\title{convert units to a chosen value}
\usage{
\method{convert}{pmunits}(x, .to, .from = attr(x, "pmunits"))
}
\arguments{
\item{x}{unit column}

\item{.to}{units to convert to (can be string, or a vector with the first element the units and the second
element named "MW" containing the moleclar weight in g/mol)}
}
\description{
convert units to a chosen value
}
