% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_units_from_list}
\alias{convert_units_from_list}
\title{convert_units_from_list converts units for all output columns provided}
\usage{
convert_units_from_list(.df, .ul)
}
\arguments{
\item{.df}{dataframe}

\item{.ul}{unit list with name/value}
}
\description{
convert_units_from_list converts units for all output columns provided
}
\examples{
library(dplyr)
Theoph <- set_units_from_list(Theoph, list(Wt = "kg", conc = "mg/L", Dose = "mg/kg"))
Theoph <- Theoph \%>\% mutate(Amt = Dose*Wt)
convert_units_from_list(Theoph, list(Amt = "g"))
}
