% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{count_if}
\alias{count_if}
\title{Increments a counter when the criteria is met}
\usage{
count_if(criteria, fill.na = T)
}
\arguments{
\item{criteria}{A logical expression that is TRUE when the count should increment}

\item{fill.na}{TRUE if NA values in criteria should be filled with FALSE or FALSE to leave them as NA.}
}
\value{
A vector with the running count of times the criteria has been met.
}
\description{
Increments a counter when the criteria is met
}
\examples{
EVID=c(1,1,0,0,1,1,1,0)
CMT= c(1,2,1,2,1,2,NA,2)
count_if(EVID==1 & CMT==1)
count_if(EVID==1 & CMT==1, fill.na=FALSE)
}
