% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{covariate_plots}
\alias{covariate_plots}
\title{Summary plot of dataset}
\usage{
covariate_plots(.data, cols = "", grouping = "")
}
\arguments{
\item{.data}{NONMEM dataset}

\item{cols}{character vector containing columns to summarise}

\item{grouping}{character vector containing columns to group by}
}
\value{
a dataframe
}
\description{
Summary plot of dataset
}
