% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_clinical.R
\name{crcl}
\alias{crcl}
\title{Compute creatinine clearance}
\usage{
crcl(age, wt, female, creat, method = c("cg"), units = "ml/min")
}
\arguments{
\item{age}{Age in years}

\item{wt}{Body mass in kg}

\item{female}{TRUE if female, FALSE if male}

\item{creat}{Serum creatinine in mg/dL}

\item{method}{"cg" for cockroft-gault is default, and only, method.}

\item{units}{Desired output units.  ml/min is default and only option.}
}
\value{
vector of computed crcl values
}
\description{
Compute creatinine clearance
}
\examples{
crcl(45, 83, TRUE, .5)
}
