% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{createDDD}
\alias{createDDD}
\title{Generate Data Definition Table}
\usage{
createDDD(DM, .data)
}
\arguments{
\item{DM}{A data management object}

\item{.data}{A NONMEM database (not yet formatted for NONMEM)}
}
\value{
A tbl object with columns for Column name, Type, Codes, Mapping, Source
}
\description{
Generate Data Definition Table
}
