% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{DataManagement}
\alias{DataManagement}
\title{Create a Data Management Object}
\usage{
DataManagement(settings_file)
}
\arguments{
\item{settings_file}{The YAML settings file to use}
}
\value{
A DataManagement object
}
\description{
Create a Data Management Object
}
