% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{date_in_range}
\alias{date_in_range}
\title{Test where a date falls relative to a date range}
\usage{
date_in_range(target, start, end, .before = FALSE, .ongoing = TRUE,
  .during = TRUE, .after = FALSE)
}
\arguments{
\item{target}{The date of interest}

\item{start}{The starting date of the range}

\item{end}{The ending date of the range}

\item{.before}{The value to assign if the date falls before the range (default FALSE)}

\item{.ongoing}{The value to assign if the date falls after start, and end is NA (default TRUE)}

\item{.during}{The value to assign if the date falls within the range (default TRUE)}

\item{.after}{The value to assign if the date falls after the range (default FALSE)}
}
\value{
The return type is determined by the user.  The default is logical.
}
\description{
Test where a date falls relative to a date range
}
\details{
The return values can be logical, character, or numeric.  The interval for within includes
the start and end dates.
}
\examples{
date_in_range(iso_to_posix("2018-03-14T12:34"), iso_to_posix("2018-03-11T12:34"), iso_to_posix("2018-03-18T12:34"))
date_in_range("2018-03-14", "2018-03", "2018-03-18T12:34")
date_in_range(iso_to_posix("2018-03-14T12:34"), iso_to_posix("2018-03-15T12:34"), iso_to_posix("2018-03-18T12:34"))
date_in_range(iso_to_posix("2018-03-14T12:34")+(1:10)*24*2600,
  iso_to_posix("2018-03-11T12:34"),
  iso_to_posix("2018-03-18T12:34"))
date_in_range(iso_to_posix("2018-03-09T12:34")+(1:10)*24*3600,
  iso_to_posix("2018-03-11T12:34"),
  c(rep(iso_to_posix("2018-03-15T12:34"),9), NA),
  .before="BEFORE", .ongoing="ONGOING", .during="DURING", .after="AFTER")
}
