% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{format_date}
\alias{format_date}
\alias{format_time}
\title{Date and Time formatting for text output}
\usage{
format_date(x, format = "\%m/\%d/\%Y", tz = "UTC")

format_time(x, format = "\%H:\%M:\%S", tz = "UTC")
}
\arguments{
\item{x}{POSIXct values}

\item{format}{A format string. see \code{base::format}.}

\item{tz}{A time zone string, defaults to UTC (Univeral)}
}
\value{
The date or time formatted as text
}
\description{
Date and Time formatting for text output
}
\details{
These functions simply provide a shorthand for the most common formats.
}
\examples{
format_date(Sys.time())
format_date(Sys.time(), format="\%Y-\%m-\%d T \%H:\%M")
format_time(Sys.time())
}
