% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_column.R
\name{default_missing_column}
\alias{default_missing_column}
\title{Create a default value in a dataframe column if the column doesn't already exist}
\usage{
default_missing_column(column, default_value = NA, env = dynGet("."))
}
\arguments{
\item{column}{string name for column to default}

\item{default_value}{default value for column, of length 1 or number of rows of dataframe}

\item{env}{the environment (dataframe) in which to evaluate the column}
}
\value{
the values of either the existing column, or the default_value if the column does not exist
}
\description{
Create a default value in a dataframe column if the column doesn't already exist
}
\details{
This function is designed to work from within a dplyr mutate call.
}
\examples{

library(dplyr)
mtcars \%>\% as_data_frame() \%>\% mutate(newcol=default_missing_column("newcol", 1))
mtcars \%>\% as_data_frame() \%>\% mutate(newcol=default_missing_column("newcol", 1:32))

}
