% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics.domain}
\alias{diagnostics.domain}
\title{## Diagnostics
## diagnostic functions should return a list of objects to be printed.  For example, a list of character messages.
##   or a data.frame}
\usage{
diagnostics.domain(dom)
}
\arguments{
\item{dom}{}
}
\value{
a list of diagnostics.
}
\description{
Run diagnostics on a loaded domain object
}
