% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_premerge}
\alias{diagnostics_premerge}
\title{Pre-merge diagnostic checks}
\usage{
diagnostics_premerge(DM)
}
\arguments{
\item{DM}{}
}
\value{
A list of diagnostic messages
}
\description{
These checks occur after the domains are loaded and before the queries are processed.  The goal is to inform on
problems that may be found before attempting to run the query processing and merging.
}
