% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\title{Create a domain object from a list of settings.}
\usage{
domain(dom.settings)
}
\arguments{
\item{dom.settings}{The settings}
}
\value{
A S3 object of class 'domain'
}
\description{
Create a domain object from a list of settings.
}
\details{
Settings may come from a yaml settings file, but must
  be converted to a list before passing in.
}
