% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domains.save.metadata}
\alias{domains.save.metadata}
\title{Save meta data for a list of domains}
\usage{
domains.save.metadata(dom.l, filename)
}
\arguments{
\item{dom.l}{A list of domains for which to save metadata}

\item{filename}{A file into which to write the YAML metadata}
}
\description{
Save meta data for a list of domains
}
