% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_fill.R
\name{.}
\alias{.}
\title{Generate a character vector from unquoted names}
\usage{
.(...)
}
\arguments{
\item{...}{names to turn into text and place in a character vector}
}
\value{
A character vector of names
}
\description{
Generate a character vector from unquoted names
}
\details{
This function is particularly useful when typing, for instance, a list of column names.
\code{.(ID,TIME)} is the same as \code{c("ID","TIME")}
}
\examples{
.(this,is,how,it,works)

}
