% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{elapsed.time}
\alias{elapsed.time}
\title{Compute elapsed time from beginning of POSIX vector}
\usage{
elapsed.time(x, units = "hours")
}
\arguments{
\item{x}{A vector of date/times POSIX objects}

\item{units}{The output units for the time difference}
}
\value{
The time difference for each member of x from the first member of x.
}
\description{
Compute elapsed time from beginning of POSIX vector
}
\details{
The first member of x will have an elapsed time of 0.
Units can be specified as per \link[base]{difftime}.  Times are not sorted here,
so if they are not in order, or at least having the minimum value first, they should
be sorted beforehand.
}
\examples{
x = Sys.time() + c(0:10)*3600
as.numeric(elapsed.time(x))

}
