% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_text.R
\name{enumerate}
\alias{enumerate}
\title{Give numeric identifiers to a vector of values}
\usage{
enumerate(vals)
}
\arguments{
\item{vals}{The values to enumerate}
}
\value{
the enumeration of vals
}
\description{
Give numeric identifiers to a vector of values
}
\details{
Given a vector of some type, assigns a number to each unique value and returns the
vector of the values' positions in the unique values.  The unique values are tabulated in
the order of appearance.
}
\examples{
enumerate(letters[1:10])
enumerate(rep(letters[1:3], times=5))
}
