% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_fill.R
\name{fill_NA}
\alias{fill_NA}
\alias{fill_locf}
\alias{fill_nocb}
\title{Fill NA values in columns}
\usage{
fill_NA(.data, column, value, groups = "ID")

fill_locf(.data, column, groups = "ID", tolerance = Inf,
  time = "TIME", backfill = T)

fill_nocb(.data, column, groups = "ID", tolerance = Inf,
  time = "TIME", backfill = T)
}
\arguments{
\item{.data}{A data.frame in which to modify the column}

\item{column}{The column to modify as an unquoted name}

\item{value}{A value or expression to use in replacing NA values}

\item{groups}{A character vector of group column names}

\item{tolerance}{A numeric value for the tolerance of the supplied time column.  Defaults to Inf.}

\item{time}{The column in .data (as a quoted string) that is used to determine if too much time has passed to carry
values.}

\item{backfill}{A logical value indicating whether NA values at the beginning of the vector
should be filled backward from the next value. (for fill_nocb, fills terminal NA values from last
value.)  Defaults to TRUE.}
}
\value{
the modified data.frame
}
\description{
Fill NA values in columns
}
\details{
The fill_* functions update NA values in a data.frame column.
\code{fill_NA} updates all NA values with the supplied value or expression
\code{fill_locf} fills NA forward unless the difference in \code{time} exceeds \code{tolerance}.
\code{fill_nocb} fills NA backward unless the difference in \code{time} exceeds \code{tolerance}.
Both locf and nocb fills fill to the start and end (respectively) of the column.  That is, for locf, if the
first values are NA, they are filled back from the next non-NA value.  Similare behavior is applied at the end of
column in the case of nocb, when the final values are NA.
}
