% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_fill.R
\name{fill_occasion}
\alias{fill_occasion}
\title{Fills occasion in a dataframe}
\usage{
fill_occasion(.data, .col, x1, x2)
}
\arguments{
\item{.data}{A dataframe}

\item{.col}{The column to create/modify}

\item{x1}{The initial condition for the occasion}

\item{x2}{The second condition for the occasion}
}
\value{
a dataframe
}
\description{
Fills occasion in a dataframe
}
\details{
See occasion
}
