% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{find_subjects_no_obs}
\alias{find_subjects_no_obs}
\title{Find subjects with no observation records}
\usage{
find_subjects_no_obs(.data)
}
\arguments{
\item{.data}{a data.frame or tbl object}
}
\value{
a single column tbl with unique IDs for subjects with no observations
}
\description{
Find subjects with no observation records
}
