% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{find_subjects_placebo_obs}
\alias{find_subjects_placebo_obs}
\title{Find placebo subjects with positive DV}
\usage{
find_subjects_placebo_obs(.data, is.placebo, condition = "DV>0")
}
\arguments{
\item{.data}{a data.frame or tbl object}

\item{is.placebo}{An R expression, as character, that is used to determine if a subject is a placebo subject.}

\item{condition}{An R expression, as character, that is tested for the placebo subjects.}
}
\value{
a single column tbl with unique IDs for placebo subjects meeting the criteria
}
\description{
The data is filtered according to is.placebo & condition, along with EVID==0 (to only query observations) and the
unique subject IDs satisfying all conditions are returned.  Condition may be a compound test, such as
"DV>0 & CMT==2".
}
