% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{formatDDD}
\alias{formatDDD}
\title{Format a DDD object as a dataframe}
\usage{
formatDDD(ddd, type = "md")
}
\arguments{
\item{ddd}{A DDD data frame}

\item{type}{md for markddown formatting of multiline cells}
}
\value{
a data_frame (tbl_df) object
}
\description{
Format a DDD object as a dataframe
}
