% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addl_dosing.R
\name{get_addl_dosing}
\alias{get_addl_dosing}
\title{derive necessary information about addl and append to given dataframe}
\usage{
get_addl_dosing(df, .f = roundup, .tolII = 0.5)
}
\arguments{
\item{df}{dataframe}

\item{.f}{function to calculate how to calculate addl for non-nominal times}

\item{.tolII}{A number between 0 and 1 that determines the fraction
of the dosing interval that must pass before an ADDL sequence can begin.}
}
\description{
derive necessary information about addl and append to given dataframe
}
