% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\name{get_baseline}
\alias{get_baseline}
\title{Find a baseline value}
\usage{
get_baseline(col, .filter, ..., fun.summ = mean_)
}
\arguments{
\item{col}{An unquoted variable name}

\item{.filter}{An unquoted R logical expression (evaluates to TRUE or FALSE).  Pass TRUE to select
all records for summarization.}

\item{...}{Comma separated, unquoted, additional grouping columns}

\item{fun.summ}{An unquoted function name to use to summarise multiple baseline values per group}
}
\value{
Returns a vector of selected baseline values
}
\description{
Find a baseline value
}
\details{
get_baseline can be used inside a mutate or summarise call, and therefore inside the get mapped
  value functions (e.g., \code{getDV}). It finds the rows that satisfy the criteria in the .filter argument, then applies fun.summ
  (a summary function) to those rows.  In this way, multiple values can be used for a baseline value.
}
\examples{
library(dplyr)
library(PMDatR)
Theoph \%>\% mutate(conc.bl = get_baseline(conc, Time>0 & Time<1, Subject, fun.summ=mean_))
Theoph \%>\% mutate(conc=set_units(conc,"mg/ml"),
  conc.bl = get_baseline(conc, Time>0 & Time<1, Subject, fun.summ=mean_)) \%>\% as_data_frame
}
