% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_domain.R
\name{get_from_domain}
\alias{get_from_domain}
\title{Get a column from a domain}
\usage{
get_from_domain(domain, column, .unique = TRUE, .unlist = TRUE)
}
\arguments{
\item{domain}{Quoted name of the domain data object}

\item{column}{Quoted name of the column from which to pull values}

\item{.unique}{A logical (default TRUE) stating if the return values should be reduced
to just unique values.}

\item{.unlist}{A logical (default TRUE) stating if the return value should be converted
an array (e.g., instead of a data_frame column)}
}
\value{
A data_frame column or array of values
}
\description{
Get a column from a domain
}
\examples{
\dontrun{
# get the unique subject ids in DM
get_from_domain("DM","USUBJID")
}

}
