% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.r
\name{get_metadata}
\alias{get_metadata}
\title{Gather metadata from a data frame}
\usage{
get_metadata(df)
}
\arguments{
\item{df}{An object derived from data.frame}
}
\value{
A metadata S3 object
}
\description{
Gather metadata from a data frame
}
