% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{get_nm_attributes}
\alias{get_nm_attributes}
\title{Get column attributes for nonmem formatted output}
\usage{
get_nm_attributes(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a dataframe
}
\description{
Get column attributes for nonmem formatted output
}
