% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MappedColumns.R
\name{getCovT}
\alias{getCovT}
\title{Processes domain data for time-varying covariates}
\usage{
getCovT(.data, ID, TIME, EVID, covT.col, covT.val, covT.filter, fun.summary,
  Units, ...)
}
\arguments{
\item{.data}{The domain data frame}

\item{ID}{A column name or expression for the subject identifier.}

\item{TIME}{A column name or expression for the TIME column.}

\item{EVID}{A column name or expression for the EVID column.}

\item{covT.col}{The name of the column that holds the desired covariate names}

\item{covT.val}{The name of the column that holds the covariate values}

\item{covT.filter}{Filter expression to apply to subset the data.}

\item{fun.summary}{A function name (or named list/vector) as a text string to use if more than one value per (ID, TIME) is encountered.}

\item{Units}{The name of the column holding the covariate units (if stacked).}

\item{...}{Additional column mappings.}
}
\value{
A data_frame with the mapped data
}
\description{
Processes domain data for time-varying covariates
}
\details{
The input data is filtered, if a filter expression is provided.  ID and TIME must be specified, but other arguments are optional.
If EVID is not provided, it will default to '2'.  Provide cov.col and cov.val both to indicate that the data are in long format.  In this case the values in cov.col
(that remain after the filter is applied) will become
new columns with the values specified in cov.val (a data widening operation).  The ID and TIME columns will be repeated
for each covariate value row.

If covT.col and covT.val are not provided, then after filtering the expressions given as additional column mappings will
be selected from .data

This function uses non-standard evaluation, so the ... arguments do not need to be quoted.  For example, to compute
a date from ISO-8601 strings include \code{TIME=parsedate::parse_date(EXSTDTC)}.  Of course the actual column may
have a different name.  Extra arguments (...) must be named, for example: \code{MYVALUE = pmax(0,DV)}.
}
