% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MappedColumns.R
\name{getDomain}
\alias{getDomain}
\title{Processes source data}
\usage{
getDomain(.data, .filter, ...)
}
\arguments{
\item{.data}{The domain data frame}

\item{.filter}{Filter expression to apply to subset the data.}

\item{...}{Additional column mappings.}
}
\value{
A data_frame with the mapped data
}
\description{
Processes source data
}
\details{
The input data is filtered, if a filter expression is provided. Named parameters provided
in the extra arguments are evaluated and selected from the input dataset.

This function uses non-standard evaluation, so the ... arguments do not need to be quoted.  For example, to compute
a date from ISO-8601 strings include \code{TIME=parsedate::parse_date(EXSTDTC)}.  Of course the actual column may
have a different name.  Extra arguments (...) must be named, for example: \code{MYVALUE = pmax(0,DV)}.
}
