% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MappedColumns.R
\name{getIndividualDoses}
\alias{getIndividualDoses}
\title{Processes domain data for individual doses}
\usage{
getIndividualDoses(.data, ID, TIME, AMT, EVID, MDV, ex.filter, Units, ...)
}
\arguments{
\item{.data}{The domain data frame}

\item{ID}{A column name or expression for the subject identifier.}

\item{TIME}{A column name or expression for the TIME column.}

\item{AMT}{A column name or expression for the AMT column.}

\item{EVID}{A column name or expression for the EVID column (defaults to 1).}

\item{MDV}{A column name or expression for the MDV column (defaults to 1).}

\item{ex.filter}{Filter expression to apply to subset the data.}

\item{Units}{A column name or expression containing the units of AMT.}

\item{...}{Additional column mappings.}
}
\value{
A data_frame with the mapped data
}
\description{
Processes domain data for individual doses
}
\details{
The input data is filtered, if a filter expression is provided.  ID, TIME, and AMT are required.  EVID
can be provided as a number or expression, but will default to 1 if not supplied.  Additional column mappings
will be added to the returned data.

This function uses non-standard evaluation, so the arguments do not need to be quoted.  For example, to compute
a date from ISO-8601 strings include \code{TIME=parsedate::parse_date(EXSTDTC)}.  Of course the actual column may
have a different name.  Extra arguments (...) must be named, for example: MYVALUE = pmax(0,DV).

AMT units may be supplied as a column reference (eg, EXDOSU) or as a fixed string (eg, "mg").  If only a single
unit is represented then AMT will be converted to a pmunits column, where the units are stored as a column attribute.
Otherweise a column AMTU will be created and carried through to the output.
}
