% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dosing_sequence.R
\name{guess_dosing_sequence}
\alias{guess_dosing_sequence}
\title{guess dosing sequence based on time of day}
\usage{
guess_dosing_sequence(.time_hrs, .ii, seq_break_times = NULL)
}
\arguments{
\item{.time_hrs}{time of day, in hours}

\item{seq_break_times}{list of time cutpoints, in hours, to use as breakpoints for given sequences}

\item{.freq}{dosing frequency (QD, BID, etc)}
}
\description{
guess dosing sequence based on time of day
}
\details{
seq_break_times determines how to calculate the sequence
defaults are as follows:
 seq_break_times <- list(
   "QD" = 25, # so always be 1
   "BID" = 12, # before noon = seq 1, after noon = seq 2
   "TID" = c(10, 16)
 )
}
