% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_fn}
\alias{impute_fn}
\title{impute a value to all missing}
\usage{
impute_fn(.df, .col, .fn, groups = "ID")
}
\arguments{
\item{.df}{data frame}

\item{.col}{column}

\item{.fn}{impute function, provides column as x}

\item{groups}{groups, defaults to ID}
}
\description{
impute a value to all missing
}
