% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_value}
\alias{impute_value}
\title{impute by a calculated value for any NA's in a column
impute a value to all missing}
\usage{
impute_value(.df, .col, .value)
}
\arguments{
\item{.df}{data frame}

\item{.col}{column}

\item{.value}{value}
}
\description{
impute by a calculated value for any NA's in a column
impute a value to all missing
}
