% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominal_time_fns.R
\name{increment_to_nominal_time}
\alias{increment_to_nominal_time}
\title{increment a time record to some nominal time}
\usage{
increment_to_nominal_time()
}
\description{
increment a time record to some nominal time
}
\details{
For dosing records with missing records, need to increment some nominal amount
}
