% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{is.number}
\alias{is.number}
\title{Determine if a value can be coerced to a numeric type}
\usage{
is.number(x)
}
\arguments{
\item{x}{A value to test.  May be an vector.}
}
\value{
A vector of same size as x or T, if the corresponding value can be coerced, or F.
}
\description{
Determine if a value can be coerced to a numeric type
}
\examples{
is.number(c(NA,"A","12"))
}
