% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_utilities.R
\name{is_file_writable}
\alias{is_file_writable}
\title{Check if a file is writable}
\usage{
is_file_writable(fname)
}
\arguments{
\item{fname}{The path to the file}
}
\value{
logical TRUE if the file is writable, FALSE if not
}
\description{
This function can test if a file is writable.  A file will not be writable if another application has it open
and locked, or if the file permissions don't support writing
}
