% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{load.domain.mappings}
\alias{load.domain.mappings}
\title{Load mappings list from mappings file}
\usage{
load.domain.mappings(filename)
}
\arguments{
\item{filename}{}
}
\value{
A named list of domain InputMappings
}
\description{
Load mappings list from mappings file
}
