% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{load.domain}
\alias{load.domain}
\title{Load a domain from a domain object or from settings}
\usage{
load.domain(x, settings, .fun, .hook)
}
\arguments{
\item{x}{A domain object with a valid name and filename}

\item{settings}{Additional settings to pass to the loading function.}

\item{.fun}{An optional mapping function to process the domain after loading
for formatted text files.  If missing the function will be constructed
from the domain settings.}

\item{.hook}{An optional function to process the loaded source after
the pre-merge process (if applicable) and before calling .fun.  If missing
the function will be constructed from the domain settings.}
}
\value{
An S3 object of class domain
}
\description{
Load a domain from a domain object or from settings
}
\details{
The domain source files are loaded and additional processing
is run.  First the main source is loaded.  Then, if a pre-merge file
is indicated, the pre-merge source is loaded and the pre-merge proceeds.
After the premerge the hook function is run, which allows for arbitrary
processing of the loaded data.  Metadata (inMeta) is captured at this point.
Then the mapping function (.fun) is run and final metadata are captured.
}
