% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel.R
\name{load_excel}
\alias{load_excel}
\title{Load an Excel sheet from a workbook}
\usage{
load_excel(obj)
}
\arguments{
\item{obj}{A domain object.}
}
\value{
A tbl_df object containing the loaded data
}
\description{
Load an Excel sheet from a workbook
}
\details{
Reads a file in xls or xlsx format and casts it as a tbl_df (dplyr) object.  It must be passed a valid
domain object, that may also contain an xlSettings object.  For full description of xlSettings options see the arguments list for
\link[readxl]{read_excel}
}
