% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sas7bdat.R
\name{load_sas7bdat}
\alias{load_sas7bdat}
\title{Load a sas7bdat file into a domain object}
\usage{
load_sas7bdat(path)
}
\arguments{
\item{path}{The fully qualified filename}
}
\value{
A tbl_df object containing the loaded data
}
\description{
Load a sas7bdat file into a domain object
}
