% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_table.R
\name{load_table}
\alias{load_table}
\title{Load a delimited text file}
\usage{
load_table(obj)
}
\arguments{
\item{obj}{A domain object.}
}
\value{
A tbl_df object containing the loaded data
}
\description{
Load a delimited text file
}
\details{
Read a file in table format and casts it as a tbl_df (dplyr) object.  It must be passed a valid
domain object, that may also contain a FileSettings object.  The FileSettings object must then contain at least one
named item to specify the seperator (sep). If FileSettings is missing default options are used per \link[utils]{read.csv}.
For full description of FileSettings options see the arguments list for
\link[utils]{read.table}
}
