% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_xpt.R
\name{load_xpt}
\alias{load_xpt}
\title{Load a sas xport file into a domain object}
\usage{
load_xpt(path, tbl = 1)
}
\arguments{
\item{path}{The fully qualified filename}

\item{tbl}{If more than one table is present, which table to load}
}
\value{
a data.frame compatible object
}
\description{
Load a sas xport file into a domain object
}
