% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{locf}
\alias{locf}
\title{Fill NA values in a vector using previous non-NA value}
\usage{
locf(x, na = NA, cond = T, backfill = T)
}
\arguments{
\item{x}{A vector of values that may have NA values to fill}

\item{na}{A scalar or vector of values to treat as NA for purposes of filling}

\item{cond}{A logical scalar or vector.  NA will not be replaced if the corresponding condition is F.}

\item{backfill}{A logical indicating if initial NA values should be backfilled from the first non-NA value.}
}
\value{
The x vector with NA values replaced by previous non-NA value
}
\description{
Fill NA values in a vector using previous non-NA value
}
\details{
This is an interpolation method called Last Observation Carried Forward.
It fills in a missing or NA value with the previous non-NA value.  If backfill=TRUE (default), any initial NA
values will be filled backward from the first non-NA value.  Multiple values can
be provided to consider as a match for NA.
}
\examples{
x = c(1,"",2,3,"",".")
locf(x, na=c("","."))
locf(x, na=c("","."), cond=c(TRUE,TRUE,TRUE,TRUE,FALSE,TRUE))
locf(c(NA,NA,1,2,NA,NA,3))
locf(c(NA,NA,1,2,NA,NA,3), backfill=FALSE)
}
