% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_domain.R
\name{lookup_in_domain}
\alias{lookup_in_domain}
\title{Lookup a value in domain data}
\usage{
lookup_in_domain(domain, value_col, key_col, key_val)
}
\arguments{
\item{domain}{Quoted name of the domain data object}

\item{value_col}{Quoted name of the column from which to pull values}

\item{key_col}{Quoted name of the column to match with key_val}

\item{key_val}{An expression that evaluates to values to match in key_col}
}
\value{
a vector of values from value_col, of length of key_val
}
\description{
Lookup a value in domain data
}
\details{
This function allows one to lookup values in domain data.  The domain name, and the
two columns to use in it are given as quoted strings.  This allows the function to be used
within query mappings, which would otherwise error on the domain and column names that do
not match the query's mapped domain's columns.  The key_val is given unquoted so the expression
will be evaluated when the function is called.

One particular use is to find the treatment start date for a given ID in a query.
}
\examples{
\dontrun{
lookup_in_domain("DM", "RFXSTDTC", "USUBJID", ID)
}
}
