% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_text.R
\name{numerize}
\alias{numerize}
\title{Combine several numbers together into a string}
\usage{
numerize(..., sep = "")
}
\arguments{
\item{...}{expressions containing numerical values to concatenate}

\item{sep}{An optional separator to insert between the values in ...}
}
\value{
A character vector with the concatenated values in ...
}
\description{
Combine several numbers together into a string
}
\details{
Any number of values can be concatenated.  Numerize is
meant to work with number-like data, but will only warn if the
expressions return non-integer values.  The returned values are character
representations of (ideally) numbers, but if non-integer values are included
they will still appear.
}
\examples{
numerize(as.character(1:10), 11:20, sep=".")
numerize(as.character(1:10), letters[1:10], sep=".")
numerize(as.character(1:10), 11:20/3, sep=".")
}
