% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{OutputDataset}
\alias{OutputDataset}
\title{Format a dataframe for output to nonmem}
\usage{
OutputDataset(df, base_name, opts = NULL, column_info = NULL)
}
\arguments{
\item{df}{A tbl object}

\item{base_name}{The filename of the output csv}

\item{opts}{A list with options for digits and na}
}
\value{
A tbl object, df formatted for output to nonmem.
}
\description{
Format a dataframe for output to nonmem
}
