% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_dosing.R
\name{overlay_dosing}
\alias{overlay_dosing}
\title{overlay dosing records onto an existing ggplot}
\usage{
overlay_dosing(.p, .df)
}
\arguments{
\item{.p}{ggplot object}

\item{.df}{dosing dataframe including TIME and ADDL columns}
}
\description{
overlay dosing records onto an existing ggplot
}
