% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_text.R
\name{parse_usubjid}
\alias{parse_usubjid}
\title{Recombine pieces of a delimited string}
\usage{
parse_usubjid(id, pick_parts, sep = "-", max_parts = 5)
}
\arguments{
\item{id}{The USUBJID values}

\item{pick_parts}{a numeric value or vector specifying which parts to recombine and in which order}

\item{sep}{The delimiter between the parts.}

\item{max_parts}{The maximum number of parts to allow for.}
}
\value{
A vector of the recombined elements from id.
}
\description{
Recombine pieces of a delimited string
}
\details{
Intended for use in simplifying USUBJID values, this function takes a column of text USUBJIDs, splits them
  by their delimiter ("-") and allows for recombining the parts.
}
\examples{
parse_usubjid(c("1-2-3-4-5","3-28-123-12-41"),5)
# "5" "41"
parse_usubjid(c("1-2-3-4-5","3-28-123-12-41"),c(1,5))
# "15"  "341"
}
