% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_individual}
\alias{plot_individual}
\title{Plot individual subject profiles in NONMEM dataset}
\usage{
plot_individual(.data, idv = "TIME", strata = "", grouping = "",
  logY = F, showADDL = F)
}
\arguments{
\item{.data}{NONMEM data as a data.frame}

\item{idv}{A comma delimited string of variables for the x axis}

\item{grouping}{A comma delimited string of grouping variables.}

\item{strat}{A comma delimited string of stratification variables.}
}
\value{
a list of ggplot2 plot objects, one for each idv.
}
\description{
The individual plot will be formed using DV values, stratified by
variables in strat.  Add CMT to strat if there are multiple observations.
Data will be colored by combinations of variables found in grouping.
}
