% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_outlier_byCuts}
\alias{plot_outlier_byCuts}
\title{Plot outliers by cutpoints}
\usage{
plot_outlier_byCuts(.data, strata = "", idv = "TAD",
  cut2.arguments = list(g = 5))
}
\arguments{
\item{.data}{data}

\item{strata}{strata}

\item{idv}{independent variable}

\item{cut2.arguments}{cut2 arguments}
}
\value{
a ggplot2 plot object
}
\description{
Plot outliers by cutpoints
}
