% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_functions.R
\name{post.merge.refactoring}
\alias{post.merge.refactoring}
\title{Perform routine generic checks on NONMEM data and apply post-merge functions}
\usage{
post.merge.refactoring(.data, fun.transform, fun.filter, fun.exclude,
  options = NULL)
}
\arguments{
\item{.data}{The NONMEM database file}

\item{fun.transform}{A function that applies transformations}

\item{fun.filter}{A function that applies filters}

\item{fun.exclude}{A functiona that applies exclusion rules}
}
\value{
The refactored .data object
}
\description{
Perform routine generic checks on NONMEM data and apply post-merge functions
}
