% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{post_transform_message}
\alias{post_transform_message}
\alias{post_transform_start}
\alias{post_transform_end}
\title{Send a message during a piped operation}
\usage{
post_transform_message(df, msg)

post_transform_start(df, colname)

post_transform_end(df, colname)
}
\arguments{
\item{df}{a dataframe}

\item{msg}{A character value to print as a message.}

\item{colname}{the column name being processed.  This is not checked for validity.}
}
\value{
the unchanged df dataframe
}
\description{
Send a message during a piped operation
}
