% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_numeric.R
\name{power}
\alias{power}
\title{Applies a power transform y=a*x^b}
\usage{
power(x, a = 1, b = 1)
}
\arguments{
\item{x}{Values to transform}

\item{a}{The scalar parameter}

\item{b}{The power parameter}
}
\value{
the transformed values
}
\description{
Applies a power transform y=a*x^b
}
\examples{
power(1:3, a=10, b=.5)
}
