% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_functions.R
\name{pre.merge}
\alias{pre.merge}
\title{Pre-merge one domain with another}
\usage{
pre.merge(dom1, dom2, keys, .filter, jointype = "left", supp = F, ...)
}
\arguments{
\item{dom1}{The primary domain object that will contain the result}

\item{dom2}{The secondary domain object that will be merged into dom1}

\item{keys}{The columns to use for merging (character vector)}

\item{.filter}{A filter to apply to the secondary domain before the merge is completed.}

\item{jointype}{One of "left", "right", "inner", "full".}

\item{supp}{Logical indicating if the pre.merge function should treat \code{dom2} as
a CDISC SUPPQUAL domain.}

\item{...}{Additional named arguments for the columns to include from \code{dom2}  May include
tranfsormations. As of version 0.5 all columns from \code{dom2} are included.}
}
\value{
A tbl_df with the merged data
}
\description{
Pre-merge one domain with another
}
\details{
The first domain \code{dom1} is merged with a second domain \code{dom2}.  The
second domain is loaded, de novo, and the raw input data is available from the source file
for merging with the first domain.  If \code{supp} is TRUE then \code{dom2} will be
unstacked on QNAM and QVAL, which must be present in \code{dom2} or supplied in the additional
arguments. (\code{...}).
}
