% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{Query}
\alias{Query}
\alias{Query_DV}
\alias{Query_Cov}
\alias{Query_CovT}
\alias{Query_IndDose}
\title{Create a Query object}
\usage{
Query(node, reqMaps = NULL, optMaps = NULL, fn = NULL,
  validnames = NULL)

Query_DV(node, meta_data = NULL)

Query_Cov(node, meta_data = NULL)

Query_CovT(node, meta_data = NULL)

Query_IndDose(node, meta_data = NULL)
}
\arguments{
\item{node}{A list with query settings}

\item{reqMaps}{Names of required mappings as character}

\item{optMaps}{Names of optional mappings as character}

\item{fn}{Name of the query function to create}
}
\value{
A Query object with 'mappings' member as the passed node.  Errors encountered during screening are
listed in the 'errors' member.
}
\description{
Create a Query object
}
\details{
The settings list in node is screened for syntactical errors such as missing mappings, incorrectly
formatted R names, and unparseable R expressions.  Errors are collected and returned in a list.
}
\section{Functions}{
\itemize{
\item \code{Query_DV}: Create Dependent Variable Query object

\item \code{Query_Cov}: Create Static Covariate Query object

\item \code{Query_CovT}: Create Time-varying Covariate Query object

\item \code{Query_IndDose}: Create Individual Dosing Query object
}}

