% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.installation.qualification.R
\name{R.installation.qualification}
\alias{R.installation.qualification}
\title{Run installation qualification for PMDatR in current R environment}
\usage{
R.installation.qualification()
}
\value{
A list containing: IQ.message; packages.df - names of versions of packages that are out of compliance
}
\description{
Run installation qualification for PMDatR in current R environment
}
