% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_fill.R
\name{recode_values}
\alias{recode_values}
\title{Recode/Replace values in a data frame column}
\usage{
recode_values(.data, .col, ..., .not_found)
}
\arguments{
\item{.data}{A data_frame type object}

\item{.col}{The unquoted name of the column to recode}

\item{...}{Name-value pairs of expressions to use in replacement}

\item{.not_found}{If a value in .col is not in ... , the value to it replace with.}
}
\value{
A copy of .data with the modified column
}
\description{
Recode/Replace values in a data frame column
}
\details{
This function is used to change values in a column.  The name in the ... argument
can be a string or a quoted value.  See the examples.
}
\examples{
library(dplyr)
data_frame(ID=1:3,SEX=c("M","F","X")) \%>\% recode_values(SEX, "M"=1, "F"=2)
data_frame(ID=1:3,SEX=c("M","F","X")) \%>\% recode_values(SEX, "M"="Male", "F"="Female")
data_frame(ID=1:3,SEX=c("M","F","X")) \%>\% recode_values(ID, "1"=1101, "2"=1102)
data_frame(ID=1:3,SEX=c("M","F","X")) \%>\% recode_values(ID, "1"=1101, "2"=1102, .not_found=NA)
data_frame(ID=1:3,SEX=c("M","F","X")) \%>\% recode_values(ID, "1"=1101, "2"=1102, "4"="not here")
}
