% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_after_dose.R
\name{relative_time}
\alias{relative_time}
\title{relative time such that first record is at time 0}
\usage{
relative_time(.df, groups = "ID", .name = "RTIME", .name_fn = NULL,
  .time = "TIME", units = "hours", digits = 10)
}
\arguments{
\item{.df}{dataframe}

\item{groups}{grouping variables}

\item{.name}{output column name}

\item{.name_fn}{custom function for naming with formatting}

\item{.time}{time column name, defaults to TIME}

\item{units}{units time scale should be in, defaults to hours}

\item{digits}{digits to round, defaults to 2}
}
\description{
relative time such that first record is at time 0
}
