% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in a vector with supplied values}
\usage{
replace_values(x, ..., .not_found = NULL)
}
\arguments{
\item{x}{The vector in which to replace values}

\item{...}{Name and value pairs, where name is what is replaced and value is what it is replaced with.}

\item{.not_found}{Value to use for unmatched values.  If not provided, unmatched values are unchanged.}
}
\value{
A vector with the replaced values.  An attempt is made to convert the return type to numeric, if possible.
}
\description{
Replace values in a vector with supplied values
}
\examples{
replace_values(c("a","b","c"),a=1,b=2)
replace_values(c("a","b","c"),a=1,b=2, .not_found = NA)
replace_values(c("a","b","c"),a=1,b=2, .not_found = "")
replace_values(c("a","b","c"),a=1,b="hello", .not_found = "")
}
