% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_units.R
\name{resolve_units}
\alias{resolve_units}
\title{resolve unit requirements across data frames}
\usage{
resolve_units(.list)
}
\arguments{
\item{.list}{list of vectors of units}
}
\description{
resolve unit requirements across data frames
}
\details{
resolve units will determine the final combination of units across
all unit combinations provided as a list of vectors of units
}
\examples{
resolve_units(list(c(conc = "mg/L"), c(conc = "ug/L")))
resolve_units(list(c(conc = "mg/L"), c(conc = "ug/L", dose = "mg")))
}
