% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_utilities.R
\name{separate_rows}
\alias{separate_rows}
\title{Split rows based on a delimited string column}
\usage{
separate_rows(df, col, sep = ", ")
}
\arguments{
\item{df}{A dataframe}

\item{col}{The column to split on}

\item{sep}{The separator characters}
}
\value{
A dataframe
}
\description{
Split rows based on a delimited string column
}
\details{
This function splits a dataframe's rows based on delimited values in a column.  For instance,
a row named "Children" with "Tommy, Gillian, Betsy" would be duplicated three times with the children
listed separately on each row.
}
\examples{
library(dplyr)
data_frame(Parent=c("Homer, Marge", "Lorelai", "Peter, Lois"), Child=c("Bart, Lisa, Maggie",
"Rory", "Chris, Meg, Stewey")) \%>\% separate_rows("Parent") \%>\% separate_rows("Child")
}
