% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_units.R
\name{set_units}
\alias{set_units}
\title{set_units sets the units for a given vector}
\usage{
set_units(x, units, strict = F)
}
\arguments{
\item{x}{vector of values}

\item{units}{units}
}
\description{
set_units sets the units for a given vector
}
\examples{
dose <- set_units(c(100, 100), "mg/kg")
}
