% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_units.R
\name{set_units_from_list}
\alias{set_units_from_list}
\title{set_units_from_list sets units for all columns that match a name}
\usage{
set_units_from_list(.df, .ul)
}
\arguments{
\item{.df}{dataframe to set units}

\item{.ul}{unit list with name/value}
}
\description{
set_units_from_list sets units for all columns that match a name
}
\examples{
set_units_from_list(Theoph, list(Wt = "kg", conc = "mg/L", Dose = "mg/kg"))
}
