% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread_}
\alias{spread_}
\title{spread that understands units}
\usage{
spread_(data, key, value, ..., .units_col = NULL, .units_key = NULL,
  .col_precedence = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{key}{The bare (unquoted) name of a column whose values will be used as column headings}

\item{value}{The bare (unquoted) name of a column whose values will populate the cells}

\item{...}{parameters to pass to the underlying tidyr::spread function}

\item{.units_col}{column associated with units in the value column}

\item{.units_key}{a list of column names and units to attempt to set}

\item{.col_precendence}{whether units found in the units_col take precedence over those found in the key}
}
\description{
spread that understands units
}
