% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfuns.R
\name{statfuns}
\alias{statfuns}
\alias{mean_}
\alias{mean}
\alias{median_}
\alias{sum_}
\alias{min_}
\alias{max_}
\alias{first_}
\alias{last_}
\alias{nth_}
\title{Summary functions overridden by PMDatR}
\usage{
mean_(x, ..., na.rm = TRUE)

median_(x, na.rm = TRUE)

sum_(..., na.rm = TRUE)

min_(..., na.rm = TRUE)

max_(..., na.rm = TRUE)

first_(x, default = NA)

last_(x, default = NA)

nth_(x, n, default = NA)
}
\arguments{
\item{x}{An R object with a mean method (e.g. numeric/logical vectors)}

\item{...}{Additional arguuments passed to or from other methods}

\item{na.rm}{TRUE (the default) if NA values should be ignored in the calculation}

\item{n}{For 'nth' the selector to use.  Note that NAs are skipped.}
}
\value{
The value of the summary function applied to vector
}
\description{
Summary functions overridden by PMDatR
}
\details{
Some summary functions have been overridden to make them easier to use from custom function.
  All of these skip NA values.  If all values in x are NA, then these functions return NA, rather than
  other values such as -Inf, Inf, or NaN
}
\examples{
mean(c(1,NA,3)) #2
median(c(1,NA,3, NA, 5)) #3
sum(c(1,NA,3)) #4
min(c(1,NA,3, NA, 5)) #1
max(c(1,NA,3, NA, 5)) #5
}
