% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_units.R
\name{strip_units}
\alias{strip_units}
\title{strip attributes from a vector if they exist}
\usage{
strip_units(.x)
}
\arguments{
\item{.x}{vector with units}
}
\description{
strip attributes from a vector if they exist
}
