% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_units.R
\name{strip_units_df}
\alias{strip_units_df}
\title{strip all unit columns from a dataframe}
\usage{
strip_units_df(.df)
}
\arguments{
\item{.df}{dataframe to strip units attributes}
}
\description{
strip all unit columns from a dataframe
}
