% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary_table}
\alias{summary_table}
\title{Summary table of dataset}
\usage{
summary_table(.data, cols = "", grouping = "", catFn = table,
  conFn = summary)
}
\arguments{
\item{.data}{NONMEM dataset}

\item{cols}{character vector containing columns to summarise}

\item{grouping}{character vector containing columns to group by}

\item{catFn}{function to use when summarising non-numeric cols}

\item{conFn}{function to use when summarising numeric cols}
}
\value{
a dataframe
}
\description{
Summary table of dataset
}
