% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{table_outlier_byCuts}
\alias{table_outlier_byCuts}
\title{Summarise DV by strata or cutpoint to look for outliers}
\usage{
table_outlier_byCuts(.data, strata = "", cut2.arguments = list(g = 5),
  idv = "")
}
\arguments{
\item{.data}{data}

\item{strata}{A comma separated string or character vector of columns to stratify by}

\item{cut2.arguments}{Arguments to Hmisc::cut2, as a list}

\item{idv}{The column to cut by}
}
\value{
a dataframe
}
\description{
Summarises DV by strata, if provided, or by cut points of idv.
}
