% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{table_time}
\alias{table_time}
\title{Summarise time variables in NONMEM dataset}
\usage{
table_time(.data, strata = "", tad.tolerance = 674)
}
\arguments{
\item{.data}{data}

\item{tad.tolerance}{time after dose tolerance}

\item{MAP}{map}
}
\value{
A data.frame
}
\description{
Summarise time variables in NONMEM dataset
}
