% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_after_flag.R
\name{time_after_criteria}
\alias{time_after_criteria}
\title{time_after_criteria calculates the time after some filter critera within a group}
\usage{
time_after_criteria(.df, .name, ..., groups = "ID", .time = "TIME",
  .name_fn = NULL, units = "hours", digits = 2)
}
\arguments{
\item{.df}{dataframe}

\item{.name}{output column name}

\item{...}{filter criteria}

\item{groups}{grouping variables}

\item{.name_fn}{custom function for naming with formatting}

\item{units}{units time scale should be in, defaults to hours.  Possible values include: secs, mins, hours, days, weeks}

\item{digits}{digits to round, defaults to 2}

\item{.col}{column}
}
\description{
time_after_criteria calculates the time after some filter critera within a group
}
\details{
this is a very flexible function that can be used to do almost any time_from_x calculcation.
For example, to given a conmed column with admininstration flagged as "Y", time from
conmed could be calculated simply with
df %>% time_after_critiera("TACONMED", conmed == "Y")
}
