% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_after_dose.R
\name{time_after_first_dose}
\alias{time_after_first_dose}
\title{time after the first dose}
\usage{
time_after_first_dose(.df, groups = "ID", .name = "TAFD",
  .criteria = EVID \%in\% c(1, 4), .name_fn = NULL, .time = "TIME",
  units = "hours", digits = 10)
}
\arguments{
\item{.df}{dataframe}

\item{groups}{grouping variables}

\item{.name}{output column name}

\item{.criteria}{a logical expression to identify relavent doses.  Defaults to EVID equal 1 or 4 (dose records).}

\item{.name_fn}{custom function for naming with formatting}

\item{.time}{time column name, defaults to TIME}

\item{units}{units time scale should be in, defaults to hours}

\item{digits}{digits to round, defaults to 2}
}
\description{
time after the first dose
}
