% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{time_diff}
\alias{time_diff}
\title{Time difference}
\usage{
time_diff(x1, x2, units = "hours")
}
\arguments{
\item{x1}{A Posixct date/time}

\item{x2}{A Posixct date/time}

\item{units}{A string giving units of: secs, min, hours, days, weeks, months, years}
}
\value{
a pmunits vector with the time difference expressed in the chosen units
}
\description{
Time difference
}
\examples{
time_diff(iso_to_posix("2001-10-01T12:00"),iso_to_posix("2000-10-01T12:00"), units="years")
time_diff(iso_to_posix("2001-10-03T12:00"),iso_to_posix("2001-10-01T12:00"), units="days")
time_diff(c(iso_to_posix("2001-10-03T12:00"),iso_to_posix("2001-10-02T12:00")),iso_to_posix("2001-10-01T12:00"), units="hours")
}
