% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_after_dose.R
\name{time_of_first_dose}
\alias{time_of_first_dose}
\title{Time of the first dose}
\usage{
time_of_first_dose(.df, .name = "FDDTTM", .criteria = EVID \%in\% c(1,
  4), .time = "TIME")
}
\arguments{
\item{.df}{dataframe}

\item{.name}{output column name as a string}

\item{.criteria}{An expression that identifies valid dose records}

\item{.time}{time column name, as a string. Defaults to TIME}
}
\description{
Time of the first dose
}
\details{
This functions operates on the entire dataframe given by .df and finds the first time specified in
  .time that matches the expression passed in .critera (by default it looks for EVID 1 or 4).  The value is returned in
  a newly created column specified by .name.
}
