% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_cols.R
\name{unit_cols}
\alias{unit_cols}
\title{get the units columns present in a dataframe and what the units are}
\usage{
unit_cols(.df)
}
\arguments{
\item{.df}{dataframe}
}
\description{
get the units columns present in a dataframe and what the units are
}
\examples{
th1 <- set_units_from_list(Theoph, list(conc = "mg/L", Wt = "kg"))
unit_cols(th1)
}
