% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmunits.R
\name{units.pmunits}
\alias{units.pmunits}
\title{retrieve measurement units from \code{pmunits} object}
\usage{
\method{units}{pmunits}(x)
}
\description{
retrieve measurement units from \code{pmunits} object
}
