% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_dosing.R
\name{visualize_dosing}
\alias{visualize_dosing}
\title{visualize dosing history}
\usage{
visualize_dosing(.df, .size = 3)
}
\arguments{
\item{.df}{dosing dataframe}

\item{.size}{size of points}
}
\description{
visualize dosing history
}
\details{
need columns TIME
}
