% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{WriteCode.DataManagement}
\alias{WriteCode.DataManagement}
\title{Write R functions corresponding to a DataManagement objet}
\usage{
WriteCode.DataManagement(DMobj)
}
\arguments{
\item{DMobj}{A data management object}
}
\value{
Text containing the functions that can be sourced or saved to a file.
}
\description{
Write R functions corresponding to a DataManagement objet
}
