% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{find_subjects_no_dosing}
\alias{find_subjects_no_dosing}
\title{Find subjects with no dosing records}
\usage{
find_subjects_no_dosing(.data, ID = "ID")
}
\arguments{
\item{.data}{a data.frame or tbl object}

\item{ID}{a character value giving the column containing subject IDs}
}
\value{
a single column tbl with unique IDs for subjects with no dosing records
}
\description{
Find subjects with no dosing records
}
