% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_time.R
\name{iso_to_posix}
\alias{iso_to_posix}
\title{Convert an ISO8601 date/time to POSIXct}
\usage{
iso_to_posix(dtc, .time, .tz = "UTC")
}
\arguments{
\item{dtc}{A ISO8601 formatted date/time}

\item{.time}{If dtc is date-only, a time to use as an offset.  If length is not 1 or
same as dtc, .time is recycled with a warning.}

\item{.tz}{The timezone for the dtc values.  Default is UTC.}
}
\value{
date/time as R POSIXct native type.
}
\description{
Convert an ISO8601 date/time to POSIXct
}
\examples{
iso_to_posix("2016-10-03T21:22:45")
iso_to_posix(c("2016-10-03","2016-10-03T12:34"),"09:00")
iso_to_posix(c("2016-10-03","2016-10-03","2016-10-03"), c("09:00", "10:10")) #gives a warning about length of .times
}
