% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{table_dv}
\alias{table_dv}
\title{Summarise observations by EVID in NONMEM dataset}
\usage{
table_dv(.data, strata = "")
}
\arguments{
\item{.data}{data}

\item{MAP}{map}
}
\value{
A dataframe
}
\description{
Summarise observations by EVID in NONMEM dataset
}
